- moving the wordnet package to the ml-common so can be used by flink - The original test classes are updated but remain in the same place as they rely on a spark testing package
