+++ sansa-ml-spark/pom.xml
@@ -106,21 +106,6 @@
 			<version>${jena.version}</version>
 		</dependency>
 
-		<!-- WordNet Library -->
-		<dependency>
-			<groupId>net.sf.extjwnl</groupId>
-			<artifactId>extjwnl</artifactId>
-			<version>1.9.4</version>
-		</dependency>
-
-		<!-- WordNet Library Dataset -->
-		<dependency>
-			<groupId>net.sf.extjwnl</groupId>
-			<artifactId>extjwnl-data-wn31-map</artifactId>
-			<version>1.0</version>
-		</dependency>
-
-
 		<!-- BigDL Library -->
 		<dependency>
 			<groupId>com.intel.analytics.bigdl</groupId>
