+++ sansa-ml-spark/src/test/scala/net/sansa_stack/ml/spark/wordnet/TestAllHypernims.scala
@@ -1,0 +1,29 @@
+package net.sansa_stack.ml.spark.wordnet
+
+import com.holdenkarau.spark.testing.DataFrameSuiteBase
+import net.sansa_stack.ml.common.nlp.wordnet.WordNet
+import net.sf.extjwnl.data._
+import org.scalatest.FunSuite
+
+class TestAllHypernims extends FunSuite with DataFrameSuiteBase {
+
+  test("Tests getting all hypernyms of the the first synset in the word cat") {
+    try {
+      val wn = new WordNet
+      val dict = wn.getDict
+        // getting a synset by a word and index
+
+        val cat = wn.getSynset("cat", POS.NOUN, 1).head
+
+        val getAllHypers = wn.getAllHypernyms(cat)
+
+         assert(getAllHypers != null)
+      }
+      catch {
+        case e: ExceptionInInitializerError => println("The WordNet dictionary is not installed, please check the readme for instructions to enable it.")
+      }
+
+  }
+}
+
+
