+++ sansa-ml-spark/src/test/scala/net/sansa_stack/ml/spark/wordnet/TestGetMaxDepth.scala
@@ -1,0 +1,27 @@
+package net.sansa_stack.ml.spark.wordnet
+
+import com.holdenkarau.spark.testing.DataFrameSuiteBase
+import net.sansa_stack.ml.common.nlp.wordnet.WordNet
+import net.sf.extjwnl.data._
+import org.scalatest.FunSuite
+
+class TestGetMaxDepth extends FunSuite with DataFrameSuiteBase {
+
+  test("Test the function that gets the maximum depth of dataset graph ") {
+
+    try {
+      val wn = new WordNet
+      val dict = wn.getDict
+
+      val thing1 = wn.getSynset("thing", POS.NOUN, 1).head
+      val dog = wn.getSynset("dog", POS.NOUN, 1).head
+
+
+      val dogD = wn.maxDepth(dog)
+      val dogD2 = wn.minDepth(dog)
+      assert(dogD != 0)
+    } catch {
+      case e: ExceptionInInitializerError => println("The WordNet dictionary is not installed, please check the readme for instructions to enable it.")
+    }
+  }
+}
