+++ sansa-ml-spark/src/test/scala/net/sansa_stack/ml/spark/wordnet/TestGetSynsets.scala
@@ -1,10 +1,11 @@
-package net.sansa_stack.ml.spark.nlp.wordnet
+package net.sansa_stack.ml.spark.wordnet
 
 import com.holdenkarau.spark.testing.DataFrameSuiteBase
+import net.sansa_stack.ml.common.nlp.wordnet.WordNet
 import net.sf.extjwnl.data._
 import org.scalatest.FunSuite
 
-class TestGetSynsets extends FunSuite with DataFrameSuiteBase {
+class TestGetSynsets extends FunSuite with DataFrameSuiteBase{
 
   test("If The WordNet dictionary is correctly installed synsets must not be null ") {
 
@@ -13,10 +14,9 @@ class TestGetSynsets extends FunSuite with DataFrameSuiteBase {
       val dict = wn.getDict
       // getting a synset by a word and index
 
-      val thing1 = wn.getSynset("thing", POS.NOUN, 1)
+      val thing1 = wn.getSynset("thing", POS.NOUN, 1).head
 
       // getting a list of synsets by a word
-
       val thing2 = wn.getSynsets("thing", POS.NOUN).head
 
       assert(thing1 != null)
